use juju-wait for making sure deployments are ready
With the release of pure snaps for conjure-up we've added another python
dependency `juju_wait`. This allows us to monitor deployments for when hook
firing becomes idle and then moving onto the post processing steps.

This is a much better solution then counting initial agents and checking their
workload/agent status.

We need to make the ppa distributions point to users installing the new snap way
as juju-wait is not available in the ppa or any archive and is only brought in
during the snap build phase.

Signed-off-by: Adam Stokes <battlemidget@users.noreply.github.com>