Use model-config resource-tags as well
If we don't tell Juju to apply the tags to resources, new units won't be
tagged properly when the applications are scaled.  Unfortunately, Juju
isn't 100% reliable when it comes to adding the tags during boostrap, so
we still have to do it explicitly to ensure everything gets the tag
initially.  Also, specifying the tag as model-config also means that
every instance, even non-k8s ones, get the tag.  It doesn't seem to hurt
anything, though.