Reduce default flavor disk size for openstack lxd
Requiring a root disk of 20GB for every guest will cause users to
be unable to boot guests after only having launched a modest number
as the nova scheduler is looking for a compute node to have free
space that is larger than the flavor root disk size. Given
the default setup of flavors, images etc are really just examples
to get people going it seems reasonable to drop the required
root volume size to allow more guests to be created in a default
install. I have changed the flavor name to 'm1.tiny' to emphasis
that guests based on this flavor will only get very modest resources.